Finished conversion into class, added vector_2d and line_2d class for easier linear interpolation

+++ wvf_peak_info.h
@@ -10,15 +10,15 @@ class wvf_peak_info {
   wvf_peak_info();
   wvf_peak_info(int,int,int,int);
  
-  void set_pulse_maximum(int);
-  void set_pulse_maximum_timestamp(int);
-  void set_pulse_start(int);
-  void set_pulse_end(int);
+  void set_pulse_maximum(double);
+  void set_pulse_maximum_timestamp(double);
+  void set_pulse_start(double);
+  void set_pulse_end(double);
   
-  int get_pulse_maximum(void);
-  int get_pulse_maximum_timestamp(void);
-  int get_pulse_start(void);
-  int get_pulse_end(void);
+  double get_pulse_maximum(void);
+  double get_pulse_maximum_timestamp(void);
+  double get_pulse_start(void);
+  double get_pulse_end(void);
 
   void reset(void);
   
@@ -26,9 +26,9 @@ class wvf_peak_info {
 
   
  private :
-  int _pulse_maximum;
-  int _pulse_maximum_timestamp;
-  int _pulse_start, _pulse_end;
+  double _pulse_maximum;
+  double _pulse_maximum_timestamp;
+  double _pulse_start, _pulse_end;
   bool _info_validated = false;
 
 };
